move serverinfo
